requirements.txt
streamlit==1.31.0
wandb==0.16.3
pandas==2.2.0
numpy==1.26.3
plotly==5.18.0
scikit-learn==1.4.0
anthropic==0.18.1
python-dotenv==1.0.1
GitPython==3.1.41
scipy==1.12.0
